src/main/java/example/WeatherClient.java
package example;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;
import org.springframework.web.client.RestTemplate;

@Component
public class WeatherClient {

    private final RestTemplate restTemplate;

    @Autowired
    public WeatherClient(RestTemplate restTemplate) {
        this.restTemplate = restTemplate;
    }

    public WeatherResponse yesterdaysWeather() {
        return restTemplate.getForObject("yahoo", WeatherResponse.class);
    }
}
